Refactor: Move Speech-To-Text files to stt package
The following files related to Speech-To-Text functionality have been moved to a new `stt` package:
- `BuddySpeechToTextService.java`
- `SpeechToTextService.java`
- `SttListener.java`

`MainActivity.java` has been updated to reflect these import changes.

Additionally, the Gradle wrapper has been updated from version 8.11.1 to 8.13.
